# src/backend/central_preventiva/composicao/openapi.json
{
  "components": {
    "schemas": {
...
      "ProblemaExecucao": {
        "additionalProperties": false,
        "description": "Falha da operação de execução, com ocorrência, impacto e próxima ação segura.",
        "properties": {
          "codigo": {
            "description": "Código estável que identifica o tipo da falha.",
            "title": "Codigo",
            "type": "string"
          },
          "correlacao_id": {
            "description": "Identificador único desta ocorrência de falha.",
            "title": "Correlacao Id",
            "type": "string"
          },
          "impacto": {
            "description": "Efeito prático da falha para quem consultou o recurso.",
            "title": "Impacto",
            "type": "string"
          },
          "ocorrencia": {
            "description": "O que aconteceu, em português brasileiro.",
            "title": "Ocorrencia",
            "type": "string"
          },
          "proxima_acao": {
            "description": "Próxima ação segura recomendada para contornar a falha.",
            "title": "Proxima Acao",
            "type": "string"
          }
        },
        "required": [
          "codigo",
          "correlacao_id",
          "ocorrencia",
          "impacto",
          "proxima_acao"
        ],
        "title": "ProblemaExecucao",
        "type": "object"
      },
...
      "RespostaExecucao": {
        "additionalProperties": false,
        "description": "Estado, marcos e, quando aplicável, prévia do público elegível de uma execução.",
        "properties": {
          "estado": {
            "description": "Estado atual da execução.",
            "title": "Estado",
            "type": "string"
          },
          "execucao_origem_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execução terminal que originou esta nova tentativa; nula quando esta execução não é correlacionada a nenhuma outra.",
            "title": "Execucao Origem Id"
          },
          "id": {
            "description": "Identificador da execução.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "marcos": {
            "description": "Marcos de transição já persistidos, na ordem em que ocorreram.",
            "items": {
              "$ref": "#/components/schemas/RespostaMarco"
            },
            "title": "Marcos",
            "type": "array"
          },
          "publico_elegivel_previa": {
            "description": "Amostra de até 5 incluídos no público elegível; vazia fora de 'aguardando_geracao'.",
            "items": {
              "$ref": "#/components/schemas/RespostaPreviaPublico"
            },
            "title": "Publico Elegivel Previa",
            "type": "array"
          },
          "publico_elegivel_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Quantidade total de incluídos no público elegível; nulo fora de 'aguardando_geracao'.",
            "title": "Publico Elegivel Total"
          },
          "retentativas": {
            "description": "Execuções criadas como nova tentativa a partir desta; vazia quando não houve nenhuma. Cada execução mantém seu próprio histórico, sem mesclar marcos.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Retentativas",
            "type": "array"
          }
        },
        "required": [
          "id",
          "estado",
          "marcos",
          "publico_elegivel_total",
          "publico_elegivel_previa",
          "execucao_origem_id",
          "retentativas"
        ],
        "title": "RespostaExecucao",
        "type": "object"
      },
      "RespostaExecucaoAceita": {
        "additionalProperties": false,
        "description": "Ack público de uma execução criada ou já em andamento com a mesma chave.",
        "properties": {
          "execucao_id": {
            "description": "Identificador da execução criada ou já em andamento.",
            "format": "uuid",
            "title": "Execucao Id",
            "type": "string"
          }
        },
        "required": [
          "execucao_id"
        ],
        "title": "RespostaExecucaoAceita",
        "type": "object"
      },
...
      "RespostaMarco": {
        "additionalProperties": false,
        "description": "Um marco de transição já persistido da execução.",
        "properties": {
          "causa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Causa registrada, ou nula quando não houve falha.",
            "title": "Causa"
          },
          "criado_em": {
            "description": "Instante RFC 3339 em UTC em que o marco foi registrado.",
            "format": "date-time",
            "title": "Criado Em",
            "type": "string"
          },
          "marco": {
            "description": "Nome do marco de transição.",
            "title": "Marco",
            "type": "string"
          }
        },
        "required": [
          "marco",
          "causa",
          "criado_em"
        ],
        "title": "RespostaMarco",
        "type": "object"
      },
...
      "RespostaPreviaPublico": {
        "additionalProperties": false,
        "description": "Um item da amostra do público elegível formado.",
        "properties": {
          "canal": {
            "description": "Canal preferencial do segurado, no momento da avaliação.",
            "title": "Canal",
            "type": "string"
          },
          "nome_segurado": {
            "description": "Nome do segurado sintético incluído.",
            "title": "Nome Segurado",
            "type": "string"
          }
        },
        "required": [
          "nome_segurado",
          "canal"
        ],
        "title": "RespostaPreviaPublico",
        "type": "object"
      },
...
      "SolicitacaoExecucao": {
        "additionalProperties": false,
        "description": "Corpo da solicitação de início de execução: a área monitorada a avaliar.",
        "properties": {
          "area_id": {
            "description": "Identificador da área monitorada a avaliar.",
            "format": "uuid",
            "title": "Area Id",
            "type": "string"
          }
        },
        "required": [
          "area_id"
        ],
        "title": "SolicitacaoExecucao",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Disponibiliza somente os recursos implementados da Central Preventiva, com dados sintéticos e sem envio real.",
    "summary": "API local da prova de conceito educacional.",
    "title": "Central Preventiva",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/execucoes": {
      "post": {
        "description": "Inicia, de forma idempotente, a orquestração completa de coleta, avaliação de risco e avaliação de elegibilidade para a área informada, sem exigir nenhum clique manual intermediário (RUNNER-01). Exige o cabeçalho `Idempotency-Key` em toda requisição.",
        "operationId": "iniciar_execucao_api_v1_execucoes_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitacaoExecucao"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaExecucaoAceita"
                }
              }
            },
            "description": "Execução aceita e em andamento."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemaExecucao"
                }
              }
            },
            "description": "Área monitorada desconhecida."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemaExecucao"
                }
              }
            },
            "description": "Conflito de idempotência."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemaExecucao"
                }
              }
            },
            "description": "Requisição sem o cabeçalho `Idempotency-Key`."
          }
        },
        "summary": "Iniciar uma execução preventiva",
        "tags": [
          "Execução"
        ]
      }
    },
    "/api/v1/execucoes/{execucao_id}": {
      "get": {
        "description": "Devolve o estado atual, os marcos de transição já persistidos e, quando o estado é 'aguardando_geracao', a quantidade total e uma prévia do público elegível formado (RUNNER-05). Traz também a navegação entre execuções correlacionadas: a origem desta execução e as novas tentativas criadas a partir dela, sem mesclar históricos.",
        "operationId": "consultar_execucao_api_v1_execucoes__execucao_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "execucao_id",
            "required": true,
            "schema": {
              "title": "Execucao Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RespostaExecucao"
                }
              }
            },
            "description": "Execução encontrada."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemaExecucao"
                }
              }
            },
            "description": "Execução inexistente."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemaExecucao"
                }
              }
            },
            "description": "O identificador da execução não é um UUID válido."
          }
        },
        "summary": "Consultar o acompanhamento de uma execução preventiva",
        "tags": [
          "Execução"
        ]
      }
    },
...
  }
}